+++ internal/impl/mssqlserver/input_mssqlserver_cdc.go
@@ -27,17 +27,18 @@ import (
 )
 
 const (
-	fieldConnectionString         = "connection_string"
-	fieldStreamSnapshot           = "stream_snapshot"
-	fieldSnapshotMaxBatchSize     = "snapshot_max_batch_size"
-	fieldStreamBackoffInterval    = "stream_backoff_interval"
-	fieldTablesExclude            = "exclude"
-	fieldTablesInclude            = "include"
-	fieldCheckpointLimit          = "checkpoint_limit"
-	fieldCheckpointCache          = "checkpoint_cache"
-	fieldCheckpointCacheKey       = "checkpoint_cache_key"
-	fieldCheckpointCacheTableName = "checkpoint_cache_table_name"
-	fieldBatching                 = "batching"
+	fieldConnectionString          = "connection_string"
+	fieldStreamSnapshot            = "stream_snapshot"
+	fieldMaxParallelSnapshotTables = "max_parallel_snapshot_tables"
+	fieldSnapshotMaxBatchSize      = "snapshot_max_batch_size"
+	fieldStreamBackoffInterval     = "stream_backoff_interval"
+	fieldTablesExclude             = "exclude"
+	fieldTablesInclude             = "include"
+	fieldCheckpointLimit           = "checkpoint_limit"
+	fieldCheckpointCache           = "checkpoint_cache"
+	fieldCheckpointCacheKey        = "checkpoint_cache_key"
+	fieldCheckpointCacheTableName  = "checkpoint_cache_table_name"
+	fieldBatching                  = "batching"
 
 	shutdownTimeout = 5 * time.Second
 )
@@ -64,15 +65,20 @@ This input adds the following metadata fields to each message:
 
 == Permissions
 
-To use the default Microsoft SQL Server cache, the user must have permissions to create tables and stored procedures. Refer to ` + "`" + fieldCheckpointCacheTableName + "`" + ` for additional details.
+When using the default Microsoft SQL Server based cache, the Connect user requires permission to create tables and stored procedures, and the ` + "rpcn" + `  schema must already exist. Refer to ` + "`" + fieldCheckpointCacheTableName + "`" + ` for more information.
 		`).
 	Field(service.NewStringField(fieldConnectionString).
 		Description("The connection string of the Microsoft SQL Server database to connect to.").
 		Example("sqlserver://username:password@host/instance?param1=value&param2=value"),
 	).
 	Field(service.NewBoolField(fieldStreamSnapshot).
-		Description("If set to true, the connector will query all the existing data as a part of snapshot process. Otherwise, it will start from the current Log Sequence Number position."),
+		Description("If set to true, the connector will query all the existing data as a part of snapshot process. Otherwise, it will start from the current Log Sequence Number position.").
+		Example(true).
+		Default(false),
 	).
+	Field(service.NewIntField(fieldMaxParallelSnapshotTables).
+		Description("Specifies a number of tables that will be processed in parallel during the snapshot processing stage.").
+		Default(1)).
 	Field(service.NewIntField(fieldSnapshotMaxBatchSize).
 		Description("The maximum number of rows to be streamed in a single batch when taking a snapshot.").
 		Default(1000),
@@ -87,7 +93,7 @@ To use the default Microsoft SQL Server cache, the user must have permissions to
 		Optional(),
 	).
 	Field(service.NewStringField(fieldCheckpointCache).
-		Description("A https://www.docs.redpanda.com/redpanda-connect/components/caches/about[cache resource^] to use for storing the current Log Sequence Number (LSN) that has been successfully delivered, this allows Redpanda Connect to continue from that Log Sequence Number (LSN) upon restart, rather than consume the entire state of the change table.").
+		Description("A https://www.docs.redpanda.com/redpanda-connect/components/caches/about[cache resource^] to use for storing the current Log Sequence Number (LSN) that has been successfully delivered, this allows Redpanda Connect to continue from that Log Sequence Number (LSN) upon restart, rather than consume the entire state of the change table. If not set the default Microsoft SQL Server based cache will be used, see `" + fieldCheckpointCacheTableName + "` for more information.").
 		Optional(),
 	).
 	Field(service.NewStringField(fieldCheckpointCacheTableName).
@@ -123,6 +129,7 @@ type config struct {
 	streamSnapshot        bool
 	streamBackoffInterval time.Duration
 	snapshotMaxBatchSize  int
+	snapshotMaxWorkers    int
 	tablesFilter          *confx.RegexpFilter
 	lsnCache              string
 	lsnCacheKey           string
@@ -146,6 +153,7 @@ func newMSSQLServerCDCInput(conf *service.ParsedConfig, resources *service.Resou
 	var (
 		connectionString             string
 		streamSnapshot               bool
+		snapshotMaxWorkers           int
 		streamBackoffInterval        time.Duration
 		snapshotMaxBatchSize         int
 		lsnCache, lsnCacheKey        string
@@ -163,6 +171,9 @@ func newMSSQLServerCDCInput(conf *service.ParsedConfig, resources *service.Resou
 		return nil, err
 	}
 	if streamSnapshot, err = conf.FieldBool(fieldStreamSnapshot); err != nil {
+		return nil, err
+	}
+	if snapshotMaxWorkers, err = conf.FieldInt(fieldMaxParallelSnapshotTables); err != nil {
 		return nil, err
 	}
 	if snapshotMaxBatchSize, err = conf.FieldInt(fieldSnapshotMaxBatchSize); err != nil {
@@ -193,10 +204,10 @@ func newMSSQLServerCDCInput(conf *service.ParsedConfig, resources *service.Resou
 				return nil, err
 			}
 		}
-	} else {
-		if cpCacheTableName, err = conf.FieldString(fieldCheckpointCacheTableName); err != nil {
-			return nil, err
-		}
+	}
+
+	if cpCacheTableName, err = conf.FieldString(fieldCheckpointCacheTableName); err != nil {
+		return nil, err
 	}
 
 	// checkpointing
@@ -224,6 +235,7 @@ func newMSSQLServerCDCInput(conf *service.ParsedConfig, resources *service.Resou
 			connectionString:      connectionString,
 			streamSnapshot:        streamSnapshot,
 			streamBackoffInterval: streamBackoffInterval,
+			snapshotMaxWorkers:    snapshotMaxWorkers,
 			snapshotMaxBatchSize:  snapshotMaxBatchSize,
 			lsnCache:              lsnCache,
 			lsnCacheKey:           lsnCacheKey,
@@ -244,6 +256,9 @@ func newMSSQLServerCDCInput(conf *service.ParsedConfig, resources *service.Resou
 	i.publisher.cacheLSN = func(ctx context.Context, lsn replication.LSN) error {
 		return i.cacheLSN(ctx, lsn)
 	}
+
+	// Has stopped is how we notify that we're not connected. This will get reset at connection time.
+	i.stopSig.TriggerHasStopped()
 
 	batchInput, err := service.AutoRetryNacksBatchedToggled(conf, &i)
 	if err != nil {
@@ -287,49 +302,56 @@ func (i *sqlServerCDCInput) Connect(ctx context.Context) error {
 	)
 	// no cached LSN means we're not recovering from a restart
 	if i.cfg.streamSnapshot && len(cachedLSN) == 0 {
-		db, err := sql.Open("mssql", i.cfg.connectionString)
-		if err != nil {
-			return fmt.Errorf("connecting to microsoft sql server for snapshotting: %s", err)
+		if snapshotter, err = replication.NewSnapshot(i.cfg.connectionString, userTables, i.publisher, i.log, i.metrics); err != nil {
+			return fmt.Errorf("creating database snapshotter: %w", err)
 		}
-		snapshotter = replication.NewSnapshot(db, userTables, i.publisher, i.log, i.metrics)
 	} else {
 		i.log.Infof("Snapshotting disabled, skipping...")
 	}
 
 	streaming = replication.NewChangeTableStream(userTables, i.publisher, i.cfg.streamBackoffInterval, i.log)
 
-	if i.stopSig == nil {
-		i.stopSig = shutdown.NewSignaller()
-	}
+	// Reset our stop signal
+	i.stopSig = shutdown.NewSignaller()
 
 	go func() {
+		var (
+			err    error
+			maxLSN = cachedLSN
+		)
 		softCtx, _ := i.stopSig.SoftStopCtx(context.Background())
-		wg, _ := errgroup.WithContext(softCtx)
-		wg.Go(func() error {
-			var (
-				err    error
-				maxLSN = cachedLSN
-			)
 
-			// snapshot if no LSN exists then store checkpoint once complete
-			if snapshotter != nil {
-				if maxLSN, err = i.processSnapshot(softCtx, snapshotter); err != nil {
-					return fmt.Errorf("processing snapshotting: %w", err)
+		// snapshot if no LSN exists then store checkpoint once complete
+		if snapshotter != nil {
+			if maxLSN, err = i.processSnapshot(softCtx, snapshotter); err != nil {
+				if i.stopSig.IsHardStopSignalled() {
+					i.log.Errorf("Shutting down snapshotting process: %s", err)
+				} else {
+					i.log.Infof("Gracefully shutting down snapshotting process: %s", err)
 				}
-				if err := i.cacheLSN(softCtx, maxLSN); err != nil {
-					return fmt.Errorf("caching LSN after snapshotting: %w", err)
-				}
-				i.log.Debugf("Cached LSN following snapshot: '%s'", maxLSN)
+				i.stopSig.TriggerHasStopped()
+				return
 			}
+			if err = i.cacheLSN(softCtx, maxLSN); err != nil {
+				if i.stopSig.IsHardStopSignalled() {
+					i.log.Errorf("Shutting down snapshotting process: %s", err)
+				} else {
+					i.log.Infof("Gracefully shutting down snapshotting process: %s", err)
+				}
+				i.stopSig.TriggerHasStopped()
+				return
+			}
+			i.log.Debugf("Cached LSN following snapshot: '%s'", maxLSN)
+		}
 
-			// start streaming changes
-			if err := streaming.ReadChangeTables(softCtx, i.db, maxLSN); err != nil {
+		// streaming
+		wg, ctx := errgroup.WithContext(softCtx)
+		wg.Go(func() error {
+			if err := streaming.ReadChangeTables(ctx, i.db, maxLSN); err != nil {
 				return fmt.Errorf("streaming from change tables: %w", err)
 			}
-
 			return nil
 		})
-
 		if err := wg.Wait(); err != nil && !errors.Is(err, context.Canceled) {
 			i.log.Errorf("Error during Microsoft SQL Server CDC Component: %s", err)
 		} else {
@@ -410,7 +432,7 @@ func (i *sqlServerCDCInput) processSnapshot(ctx context.Context, snapshot *repli
 		_ = snapshot.Close()
 		return nil, fmt.Errorf("preparing snapshot: %w", err)
 	}
-	if err = snapshot.Read(ctx, i.cfg.snapshotMaxBatchSize); err != nil {
+	if err = snapshot.Read(ctx, i.cfg.snapshotMaxWorkers, i.cfg.snapshotMaxBatchSize); err != nil {
 		_ = snapshot.Close()
 		return nil, fmt.Errorf("reading snapshot: %w", err)
 	}